Renamed the plugin
